Redirects DayZ modules to a duplicate folder
This commit is going off of a few assumptions for the client, which are yet to be enacted:
 - A PBO must be created in the @DayZLegacy folder called and prefixed as dayz_modules. All of the contents of modules_dayz.pbo need to be duplicated into this folder.
 - Any scripts within /dzlegacy/server_data that reference the vanilla PBO should be redirected to use the modded PBO. 

This further segregates the vanilla game from the modded game, because all of the needed scripts to run the gameplay loop are duplicated in both the addons folder and my DayZLegacy folder. Ideally, one should be able to download a clean copy of 0.44 from their Steam manifests, plop in my mod and mission folder, replace DayZ.exe and launch.