Add RVExtension load prompt at spawn
This whole file has a couple of problems. It's stuck into the `clientNew` handler at `player_queued.sqf`, so every time you respawn, the scripts are recompiled. The debug menu starts to glitch with two or more instances laid over each other, and the prompt for the RVExtension enable happens every time you respawn, or sometimes doesn't appear at all, depending on server performance.

We should find a way to check if a script has already been executed, and if it has, then exit immediately.