Create "armory" in debug test area
The purpose of this "armory" is to have an instance of every weapon, item, and attachment spawned, so that they can be quality tested much easier. This is a work in progress that creates a warehouse near the test area, and spawns a number of different weapons, including the AK74, Aug, AK101, AKM, AKS-74U, Blaze 95, and CZ527. The rest of these weapons as well as their attachments will be added as time goes on, and a new warehouse will be created by the side of the other one that contains all types of loot and items. 

This commit also contains a fix for fast walking in the debug menu, and re-enables Hilux spawning from the vehicle menu.